move panel_log.py in separate directory and split by pba type
