+++ psengine/playbook_alerts/models/panel_log/code_repo.py
@@ -1,0 +1,48 @@
+##################################### TERMS OF USE ###########################################
+# The following code is provided for demonstration purpose only, and should not be used      #
+# without independent verification. Recorded Future makes no representations or warranties,  #
+# express, implied, statutory, or otherwise, regarding any aspect of this code or of the     #
+# information it may retrieve, and provides it both strictly “as-is” and without assuming    #
+# responsibility for any information it may retrieve. Recorded Future shall not be liable    #
+# for, and you assume all risk of using, the foregoing. By using this code, Customer         #
+# represents that it is solely responsible for having all necessary licenses, permissions,   #
+# rights, and/or consents to connect to third party APIs, and that it is solely responsible  #
+# for having all necessary licenses, permissions, rights, and/or consents to any data        #
+# accessed from any third party API.                                                         #
+##############################################################################################
+
+from datetime import datetime
+
+from pydantic import Field
+
+from ....common_models import IdOptionalNameType, RFBaseModel
+from .common import ChangeType, MaliciousAssessment
+
+
+class Document(RFBaseModel):
+    id_: str = Field(alias='id')
+    content: str
+    owner_id: str
+    owner_name: str | None = None
+    published: datetime
+
+
+class WatchList(RFBaseModel):
+    id_: str = Field(alias='id')
+    name: str | None = None
+
+
+class RepoAssessment(MaliciousAssessment):
+    text_indicator: str | None = None
+    entity: IdOptionalNameType | None = None
+
+
+class CodeRepoLeakageEvidence(RFBaseModel):
+    assessments: list[RepoAssessment]
+    document: Document
+    target_entities: list[IdOptionalNameType]
+    watch_lists: list[WatchList]
+
+
+class CodeRepoLeakageEvidenceChange(ChangeType):
+    added: list[CodeRepoLeakageEvidence]
