+++ psengine/playbook_alerts/models/panel_log/tpr.py
@@ -1,0 +1,43 @@
+##################################### TERMS OF USE ###########################################
+# The following code is provided for demonstration purpose only, and should not be used      #
+# without independent verification. Recorded Future makes no representations or warranties,  #
+# express, implied, statutory, or otherwise, regarding any aspect of this code or of the     #
+# information it may retrieve, and provides it both strictly “as-is” and without assuming    #
+# responsibility for any information it may retrieve. Recorded Future shall not be liable    #
+# for, and you assume all risk of using, the foregoing. By using this code, Customer         #
+# represents that it is solely responsible for having all necessary licenses, permissions,   #
+# rights, and/or consents to connect to third party APIs, and that it is solely responsible  #
+# for having all necessary licenses, permissions, rights, and/or consents to any data        #
+# accessed from any third party API.                                                         #
+##############################################################################################
+
+from datetime import datetime
+
+from ....common_models import RFBaseModel
+from .common import ChangeType
+
+
+class TPRRiskEvidence(RFBaseModel):
+    """Same fields as `Assessment` but all optional."""
+
+    level: int
+    evidence_string: str | None = None
+    timestamp: datetime | None = None
+
+
+class ThirdPartyAssessmentChange(ChangeType):
+    risk_attribute: str
+    added: TPRRiskEvidence | None = None
+    removed: TPRRiskEvidence | None = None
+
+
+class Assessment(RFBaseModel):
+    level: int
+    evidence_string: str
+    timestamp: datetime
+
+
+class AssessmentChange(ChangeType):
+    risk_attribute: str
+    removed: Assessment | None = None
+    added: Assessment | None = None
