+++ psengine/playbook_alerts/models/panel_log/vulnerability.py
@@ -1,0 +1,38 @@
+##################################### TERMS OF USE ###########################################
+# The following code is provided for demonstration purpose only, and should not be used      #
+# without independent verification. Recorded Future makes no representations or warranties,  #
+# express, implied, statutory, or otherwise, regarding any aspect of this code or of the     #
+# information it may retrieve, and provides it both strictly “as-is” and without assuming    #
+# responsibility for any information it may retrieve. Recorded Future shall not be liable    #
+# for, and you assume all risk of using, the foregoing. By using this code, Customer         #
+# represents that it is solely responsible for having all necessary licenses, permissions,   #
+# rights, and/or consents to connect to third party APIs, and that it is solely responsible  #
+# for having all necessary licenses, permissions, rights, and/or consents to any data        #
+# accessed from any third party API.                                                         #
+##############################################################################################
+
+from datetime import datetime
+
+from pydantic import Field
+
+from ....common_models import RFBaseModel
+from .common import ChangeType, MaliciousAssessment
+
+# Same shape as `MaliciousAssessment` (id/level/title); aliased to keep the public name stable
+# and avoid diverging copies.
+VulnerabilityAssessment = MaliciousAssessment
+
+
+class TriggeredRiskRule(RFBaseModel):
+    id_: str = Field(alias='id')
+    name: str | None = None
+    description: str | None = None
+    evidence_string: str | None = None
+    machine_name: str | None = None
+    timestamp: datetime | None = None
+
+
+class VulnerabilityLifecycleChange(ChangeType):
+    added: MaliciousAssessment | None = None
+    removed: MaliciousAssessment | None = None
+    triggered_by_risk_rule: TriggeredRiskRule | None = None
